requirements.txt
asgiref==3.4.1
cffi==1.15.0
cryptography==36.0.0
Django==3.2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
pycparser==2.21
pytz==2021.3
sqlparse==0.4.2
typing_extensions==4.0.0
